miniproject/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
celery
redis
drf-yasg
django-filter
Pillow
pdfkit
requests
